reworked the flow between Main and Home activities, updated /register, /login, /reset-password and /me API to include new token, so there is no need for /token-refresh API to be called unnecessarily

+++ app/src/main/java/com/example/maiplan/database/entities/AuthEntity.kt
@@ -25,7 +25,7 @@ data class AuthEntity(
 
     val username: String,
 
-    val balance: Double = 0.00,
+    val balance: Float = 0.00f,
 
     @ColumnInfo(name = "created_at")
     val createdAt: Long = System.currentTimeMillis(),
